finish moving to StatusCode rather than u16 for HTTP status code (#885)
* finish moving to StatusCode rather than u16 for HTTP status code

While trying to address #883, I found that multiple other places were
using u16 instead of StatusCode.  This updates the other areas and adds
the requested example.

* fix clippy

* remove outdated comment

* updated generated crates to use StatusCode rather than u16

+++ sdk/identity/src/client_credentials_flow/mod.rs
@@ -84,7 +84,7 @@ pub async fn perform(
     let rsp_status = rsp.status();
     let rsp_body = rsp.into_body().await;
     if !rsp_status.is_success() {
-        return Err(ErrorKind::http_response_from_body(rsp_status as u16, &rsp_body).into_error());
+        return Err(ErrorKind::http_response_from_body(rsp_status, &rsp_body).into_error());
     }
     serde_json::from_slice(&rsp_body).map_kind(ErrorKind::DataConversion)
 }
